better edition selection
